Clarify alerting category thresholds
Add more detailed information on alerting checks, their thresholds and
add first suggestions for actions to be taken if the alerts trigger.

Note for editors:

  I used the :---|:---| tables to get the page render locally. Feel free
to adjust everything to decrease the total amount of space usage on the
page, since with all separate tables and rows it's rather huge now.

  The "5.  If one of the status items is not listed" shows up as 1. now
instead of 5... Also, the whole {% markdown %} thing does not render
locally or on regular github markdown.