Update PBAR documentation after conversion (#670)
* Update PBAR documentation after conversion

* Adjust security best practices for pbar instead of request handlers
* Remove release/conversion announcements, they're in the past now
* Fix typo for top level location ' -> /

I want to leave the request-handlers-to-pbar.md for a bit. It's not
referenced from other pages now, but it's referenced from the conversion
announcement email that customers got. We can remove the page after all
dust settled down, e.g. in Oct 2017. I'll send a separate PR as a
reminder for that.

* Update environments-details.md

* Update best-practices-security-and-improvements-for-mendix-applications.md

* Update best-practices-security-and-improvements-for-mendix-applications.md