Update deploy-mendix-on-microsoft-windows.md
Based on ticket: https://mendixsupport.zendesk.com/agent/tickets/53238 and Mitya's comment (see last line):

Proper fix

If the cause is indeed such as I've guessed, then the fix would be to add cache-control header to the ngnix config for resources that are cached by the etag/date (i.e. resources without a cacheburst). Probably, example config that we give for on-premise deployments should also be corrected.