More clear and correct texts.
The documentation informed me a lot but also confused and lacked details. So I decided to improve it. First of all, there is no justification for having published-rest-service, published-rest-services and published-rest-operation. That is one page too many.  I propose to remove published-rest-service and have incorporated the information on that page in published-rest-services. Also the texts itself are now more clear and correct and will guide the developers better.