Proposal to improve readability by using smallcaps for some prefix-characters
Proposal: Make the prefixes more readable by using smallcaps for those characters that are not the first character of a word, but rather the second character or the first of one of it's syllables. Example: OCh for On Change. 
No prefix gets changed, only some of their Capitals become Smallcaps.