Update string-function-calls.md (#1471)
* Update string-function-calls.md

This issue with IsMatch('', 'whatever') is pretty weird, should not have happened or at least be documented. But since it not in the release notes, nore in the docs of communuty commons, we should add it here.

* Update string-function-calls.md

Completed my sentence.

* Update string-function-calls.md

A bit more specific now. Change 7.20.1 to 7.8.0

* Document current behavior

Cannot identify where (or if) this changed, so will just document the current status.

If we identify issues in previous versions returning the opposite value then can make a new PR.