Update custom-domains.md
Feddo and I were discussing this and the current way this text is phrased makes it sound like something Mendix does not support, while in fact it is something that it is not supported in general (e.g. https://www.isc.org/blogs/cname-at-the-apex-of-a-zone/).

We came up with the above to make this more clear and also explain how you could resolve this if you still want to use a custom apex/naked domain.