Updated Documentation for new editor layout
The Documentation editor for creating a new content is simplified,
with less options, with the hope of making it less confusing for the user.
Also modifies the editor so that one can upload images directly into the
editor. Added some documentation instructions to the user on how to do that.
Part of release 6.1.0 for App Store, expected release date of 31-7-19