Adds info about how to commit association changes.
The Add and Remove buttons don't commit the changes on their own.
It is expected that the surrounding object (the Customer data view
as shown in the docs right above this text) has a Save button.
When the user saves the changes, the association changes that were
done from the Add and Remove buttons are committed.

Since this may not be apparently obvious, I added a short note in the
docs.