Change to make clear which modeler the SDK uses
Change to make clear which modeler the SDK uses. It took me a while to figure out how to use Mendix Studio Pro for SDK, instead of the Modeler 7.23. Hopefully, with this change, it will be a little bit clearer what you need to change to be able to use other versions of the Studio Pro.