Fix Dockerfile to account for the default node user
Node docker images provides a `node` user as the non-root user to run
the docker containers.

However as we have configured `app` as the non-root user, it creates
an additional user to the container, which doesn't map well to the
local host user's uid and gid. Hence we simply modify the login and
group details of the `node` user to match the `app` user for our purposes.