More intuitive description
Trying to make the description a bit more intuitive because it is not always clear. See for instance https://forum.mendixcloud.com/link/questions/102134.
Mind you, I am not sure if a=MyApp is needed in https://myapp.xyz.com/SSO/login?a=MyApp&f=true&cont=, but i am also not sure if it is wrong, so I left that example-url unchanged.