Update mendix-sso.md
This page mentions "default administration module" in a couple of places but there is no such module inside the Mendix SSO module. I believe what is meant is that there is a Admin sub-folder which has some example pages. I have tried to change text in those two places to clarify it. Please review. I am not sure how to attach a picture but I also have a picture ready showing the Admin sub-folder if you need it.