Restructure the content with the module "Email with templates"
The module "Data Grid 2" has been shipped with Mx 9. Therefore, it shouldn't be used as an example for something that needs be downloaded from Marketplace. Having realized this, I decided to use the module "Email with templates" instead. Now, the section about how to configure the module is linked to the specific document of the module. The reasons are as follows: first, this topic should focus on how to deal with a module in general instead of how to use a specific module; second, users can just click the link and view the complete configuration information there; and third, people can see that there is a lot of useful information in the specific document of a module.