Update symbols
I tried to go as black and white as I could in accordance with our other documents' emoji. However, I could not find a B/W lock symbol for HTTPS. I am going to browse around and see if there's a solution to this that is both B/W AND combatable with all users' browsers.