SameSiteCookies 'Strict' prevents SSO to function
Some change to Custom settings to the SameSiteCookie. In the near future I may run into this when we upgrade to Mx9. This looks like a nice place to jot down this notification.
This is not my own wisdom: see forum questions 109976 and 109785. Btw. Sounds like there is likely a better solution in near future if SAML improves their inner working. For now this will have to do.