Custom HTTP header
It tooks us quite a while to figur out why a custom header was missing while the other headers did show fine. Only then did we find out that HTTP headers may not contain an underscore. Would be nice if the Mendix modeler would already check this. Now at least it is in the documentation.
Regards,
Ronald