Update debug-microflows-remotely.md
Dear Mx,

Sending this request to add a sentence on remote nanoflow debugging (which isn't possible) - the menu structure of the documentation section did suggest it is possible + the error message given in Studio Pro is not clear as to why the breakpoint is disabled. Hence suggest adding this section. Please also refer to Mx Support ticket 141362.

Best regards,

Wouter Penris