Update _index.md
For a university program I was trying out a scratch build to help the students along. But I noticed when pressing F1 that the documentation did not give me any help on what I wanted to do namely creating a date time attribuut from a string. Strange that the formatDateTime is not mentioned at all on this page while the FormatDecimal is. So hence I added it to the Date creation part. Or you should add a complete new section and also point to the formatDate and formatTime function calls because they also seem to be missing here.