Wait for search option in not explained on this page
When pressing the F1 button you end up on this page. Now the Wait for search option in not explained on this page. This option is inside the Data source part but that part also does not explain this option and imho does not belong there. The only location where the option is explained is here: https://docs.mendix.com/refguide7/database-source/ But that is for Mx7. IMHO the F1 press should point to the page https://docs.mendix.com/refguide/search-bar/ and there this option should be explained.