Update community-best-practices-for-app-performance.md
Remove microflow recommendation on 'retrieve by association'.
The statement 'Use the retrieve over association if possible. This uses caching, it is more readable, and it uses an index. If business logic requires the database value (because the value over association might be changed), then of course a database retrieve is needed.' is wrong or misleading.
There is no such caching mechanism documented anywhere and also the statement about indexes is lacking any explanation,