Update Content Security Policy documentation
Updated the Content Security Policy documentation to explain about the changes coming in 10.12. In 10.12, the header can now be configured using a custom runtime setting which offers more flexibility than setting it through the Cloud Portal (support for nonces).

It will also be the preferred approach, as it will allow developers that create custom request handlers to build on top of our implementation, instead of devising their own.