Update build-todo-example-extension.md
The page references MenuBarExtension which does not appear to be part of the API at this time (at least not that I can see in 10.12.0 or the documentation) - If you convert it to MenuExtension and modify the code a bit then it works fine. Uploading what I have done if it's of any use.